extract useAccessibilityProperties
